refactor: CardListProvider 분리
