--- src/context.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const CardListContext = React.createContext({
-  cardList: [],
-  addNewCard: () => {},
-  updateNickNameByIndex: () => {},
-});
